requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pandas>=2.0.0
python-multipart
pydantic>=2.0.0
google-generativeai>=0.3.2
python-dotenv>=1.0.0